+++ src/Panel.js
@@ -732,9 +732,6 @@ define(function (require, exports) {
         var menu = Menus.getMenu(Menus.AppMenuBar.VIEW_MENU);
         menu.addMenuDivider();
         menu.addMenuItem(PANEL_COMMAND_ID, Main.preferences.getValue("panelShortcut"));
-
-        // Make Git in status bar clickable
-        $("#git-status").addClass("clickable").on("click", toggle);
     }
 
     function enable() {