+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.8.9
 * Do not display 'not a git repo' anymore [#111](https://github.com/zaggino/brackets-git/issues/111)
+* Removed obsolete status bar [#110](https://github.com/zaggino/brackets-git/issues/110)
 
 ## 0.8.8
 * Add toggle panel keyboard shortcut in settings dialog.